+++ src/hotspin-core/common/timer.go
@@ -67,7 +67,7 @@ func (t *Timer) Average() float64 {
 }
 
 func (t *Timer) String() string {
-	return fmt.Sprint(t.Count, "\t", float32(1000*t.Average()), " ms/call")
+	return fmt.Sprint(t.Count, "\t", float64(1000*t.Average()), " ms/call")
 }
 
 // Global timers indexed by a tag string
