+++ src/hotspin-core/gpu/kappa_wrapper.go
@@ -25,18 +25,18 @@ func KappaAsync(kappa *Array, msat0 *Array, msat0T0 *Array, T *Array, Tc *Array,
 
 	// Calling the CUDA functions
 	C.kappaAsync(
-		(*C.float)(unsafe.Pointer(uintptr(kappa.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(msat0.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(msat0T0.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(T.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(Tc.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(S.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer(uintptr(n.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(kappa.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(msat0.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(msat0T0.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(T.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(Tc.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(S.Comp[X].pointer))),
+		(*C.double)(unsafe.Pointer(uintptr(n.Comp[X].pointer))),
 
-		(C.float)(msat0Mul),
-		(C.float)(msat0T0Mul),
-		(C.float)(TcMul),
-		(C.float)(SMul),
+		(C.double)(msat0Mul),
+		(C.double)(msat0T0Mul),
+		(C.double)(TcMul),
+		(C.double)(SMul),
 
 		(C.int)(kappa.partLen3D),
 		(C.CUstream)(unsafe.Pointer(uintptr(stream))))
